refactor(console): support offline theme adaption
