+++ packages/console/src/cloud/App.tsx
@@ -1,5 +1,6 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
+import { AppThemeProvider } from '@/contexts/AppThemeProvider';
 import Callback from '@cloud/pages/Callback';
 
 import * as styles from './App.module.scss';
@@ -10,14 +11,16 @@ import { CloudRoute } from './types';
 const App = () => {
   return (
     <BrowserRouter>
-      <div className={styles.app}>
-        <Routes>
-          <Route path={`/${CloudRoute.Callback}`} element={<Callback />} />
-          <Route path={`/${CloudRoute.SocialDemoCallback}`} element={<SocialDemoCallback />} />
-          <Route path={`/:tenantId/${CloudRoute.Callback}`} element={<Callback />} />
-          <Route path="/*" element={<Main />} />
-        </Routes>
-      </div>
+      <AppThemeProvider>
+        <div className={styles.app}>
+          <Routes>
+            <Route path={`/${CloudRoute.Callback}`} element={<Callback />} />
+            <Route path={`/${CloudRoute.SocialDemoCallback}`} element={<SocialDemoCallback />} />
+            <Route path={`/:tenantId/${CloudRoute.Callback}`} element={<Callback />} />
+            <Route path="/*" element={<Main />} />
+          </Routes>
+        </div>
+      </AppThemeProvider>
     </BrowserRouter>
   );
 };
