+++ packages/console/src/hooks/use-user-preferences.ts
@@ -3,8 +3,8 @@ import { useEffect, useMemo } from 'react';
 import { z } from 'zod';
 
 import { appearanceModeStorageKey } from '@/consts';
+import { getAppearanceModeFromLocalStorage } from '@/contexts/AppThemeProvider';
 import { appearanceModeGuard } from '@/types/appearance-mode';
-import { getAppearanceModeFromLocalStorage } from '@/utils/theme';
 
 import useMeCustomData from './use-me-custom-data';
